Win 10 HLS PF Plugin Port - XAML - Setup
Win 10 HLS PF Plugin Port - XAML - Setup

+++ SDK/Windows10/PlayerFramework/Microsoft.CC608/pch.h
@@ -9,4 +9,4 @@
 #include "StandardsExtensions.h"
 #include "Logger.h"
 
-typedef uint16_t byte_t;
+typedef uint8 byte_t;
